project/requirements.txt
django==3.1.2
psycopg2-binary==2.8.5
djangorestframework==3.11.0